refactor : fixture 로 추출
