+++ src/test/kotlin/com/example/estdelivery/domain/shop/ShopOwnerTest.kt
@@ -1,8 +1,7 @@
 package com.example.estdelivery.domain.shop
 
-import com.example.estdelivery.domain.coupon.Coupon
-import com.example.estdelivery.domain.coupon.CouponBook
-import com.example.estdelivery.domain.coupon.CouponType
+import com.example.estdelivery.domain.fixture.게시할_쿠폰
+import com.example.estdelivery.domain.fixture.나눠줄_쿠폰
 import com.example.estdelivery.domain.member.Member
 import com.example.estdelivery.domain.member.UnUsedCouponBook
 import io.kotest.core.spec.style.FreeSpec
@@ -18,7 +17,6 @@ class ShopOwnerTest : FreeSpec({
         단골_리스트.addRoyalCustomers(홍길동, 김철수)
 
         val 가게_주인 = ShopOwner(Shop(PublishedCouponBook(), HandOutCouponBook(), UsedCouponBook(), 단골_리스트))
-        val 나눠줄_쿠폰 = Coupon.FixDiscountCoupon(1, 1000, "1000원 할인 쿠폰", "1000원 할인 쿠폰 설명", CouponType.IS_HAND_OUT)
 
         // when
         가게_주인.handOutCouponToRoyalCustomersInShop(나눠줄_쿠폰)
@@ -29,8 +27,7 @@ class ShopOwnerTest : FreeSpec({
 
     "쿠폰을 가게에 게시한다." {
         // given
-        val 가게_주인 = ShopOwner(Shop(PublishedCouponBook(), HandOutCouponBook(),UsedCouponBook(), RoyalCustomers()))
-        val 게시할_쿠폰 = Coupon.FixDiscountCoupon(1, 1000, "1000원 할인 쿠폰", "1000원 할인 쿠폰 설명", CouponType.IS_PUBLISHED)
+        val 가게_주인 = ShopOwner(Shop(PublishedCouponBook(), HandOutCouponBook(), UsedCouponBook(), RoyalCustomers()))
 
         // when
         가게_주인.publishCouponInShop(게시할_쿠폰)
@@ -41,7 +38,7 @@ class ShopOwnerTest : FreeSpec({
 
     "단골 회원을 가게에 추가한다." {
         // given
-        val 가게_주인 = ShopOwner(Shop(PublishedCouponBook(), HandOutCouponBook(),UsedCouponBook(), RoyalCustomers()))
+        val 가게_주인 = ShopOwner(Shop(PublishedCouponBook(), HandOutCouponBook(), UsedCouponBook(), RoyalCustomers()))
         val 홍길동 = Member(1, "홍길동", UnUsedCouponBook())
         val 김철수 = Member(2, "김철수", UnUsedCouponBook())
